Fix git branch block not displaying inside of submodules
The logic for finding a Git top-level directory failed for submodules
which had a `.git` *file* present.  Instead of reinventing the wheel,
use `git rev-parse --show-toplevel` to get the top-level directory.

Similarly, use `git symbolic-ref --short HEAD` to obtain the current
branch name, and fall back to `git rev-parse --short` if that fails.